Add basic regular paste to plugin
Add the most basic functionallity to the plugin. The plugin should be
able to paste over other values. The functionallity that was currently
added to the plugin is to paste a regular copied value over a motion
from the user.

Currently, all the code in the plugin was copied from the plugin of
vim-RelaceWithRegister, and the functionallity of restoring the value of
the pasted register was removed.
This is not the best way to make the plugin work, it is just a basic POC
of the plugin, to make a basic working copy of it. This will be
refactored in the future to be written in a more appropriate manner.